Move helper to module
